Rename monkey_patch.py to patches.py and update import in apps.py
