--- src/sigic_geonode/monkey_patch_resources/tests.py
@@ -1,3 +1,0 @@
-from django.test import TestCase
-
-# Create your tests here.
